RavenDB-24021 refactored the hub by pushing down the implementation
