+++ src/Raven.Server/Documents/Replication/Outgoing/OutgoingPullReplicationHandler.cs
@@ -1,11 +1,19 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using Raven.Client.Documents.Operations.Replication;
 using Raven.Client.Documents.Replication;
 using Raven.Client.Documents.Replication.Messages;
 using Raven.Client.ServerWide.Commands;
+using Raven.Client.ServerWide.Tcp;
 using Raven.Server.Documents.Replication.Senders;
+using Raven.Server.Documents.Replication.Stats;
+using Raven.Server.ServerWide.Context;
+using Raven.Server.Utils;
+using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Logging;
+using Sparrow.Server.Utils;
+using Sparrow.Utils;
 
 namespace Raven.Server.Documents.Replication.Outgoing
 {
@@ -17,7 +25,7 @@ namespace Raven.Server.Documents.Replication.Outgoing
 
         public string CertificateThumbprint;
 
-        protected OutgoingPullReplicationHandler(ReplicationLoader parent, DocumentDatabase database, ReplicationNode node, TcpConnectionInfo connectionInfo) : 
+        protected OutgoingPullReplicationHandler(ReplicationLoader parent, DocumentDatabase database, ReplicationNode node, TcpConnectionInfo connectionInfo) :
             base(parent, database, node, connectionInfo)
         {
         }
@@ -42,10 +50,46 @@ namespace Raven.Server.Documents.Replication.Outgoing
         // we need to associate this instance to the replication definition.
         public string PullReplicationDefinitionName;
 
-        public OutgoingPullReplicationHandlerAsHub(ReplicationLoader parent, DocumentDatabase database, PullReplicationAsHub node, TcpConnectionInfo connectionInfo) : 
+        /// <summary>
+        /// The replication scope that should be disposed when the replication is done.
+        /// </summary>
+        private IDisposable _scope;
+
+        public OutgoingPullReplicationHandlerAsHub(ReplicationLoader parent, DocumentDatabase database, PullReplicationAsHub node, TcpConnectionInfo connectionInfo) :
             base(parent, database, node, connectionInfo)
         {
         }
+
+        public void StartPullReplicationAsHub(IDisposable replicationScope, Stream stream, TcpConnectionHeaderMessage.SupportedFeatures supportedVersions)
+        {
+            SupportedFeatures = supportedVersions;
+            _stream = stream;
+            _scope = replicationScope;
+            OutgoingReplicationThreadName = $"Pull replication as hub {FromToString}";
+            _longRunningSendingWork =
+                PoolOfThreads.GlobalRavenThreadPool.LongRunning(x => HandleReplicationErrors(PullReplication), null, ThreadNames.ForOutgoingReplication(OutgoingReplicationThreadName,
+                    _database.Name, Destination.FromString(), pullReplicationAsHub: true));
+        }
+
+        private void PullReplication()
+        {
+            NativeMemory.EnsureRegistered();
+
+            AddReplicationPulse(ReplicationPulseDirection.OutgoingInitiate);
+            if (Logger.IsInfoEnabled)
+                Logger.Info($"Start pull replication as hub {FromToString}");
+
+            using (_scope)
+            using (_stream)
+            using (_interruptibleRead = new InterruptibleRead<DocumentsContextPool, DocumentsOperationContext>(_parent.ContextPool, _stream))
+            using (_database.DocumentsStorage.ContextPool.AllocateOperationContext(out JsonOperationContext context))
+            using (context.GetMemoryBuffer(out _buffer))
+            {
+                InitialHandshake();
+                Replicate();
+            }
+        }
+
         public override string FromToString => $"{base.FromToString} (pull definition: {PullReplicationDefinitionName})";
     }
 
@@ -53,7 +97,7 @@ namespace Raven.Server.Documents.Replication.Outgoing
     {
         private readonly PullReplicationAsSink _node;
 
-        public OutgoingPullReplicationHandlerAsSink(ReplicationLoader parent, DocumentDatabase database, PullReplicationAsSink node, TcpConnectionInfo connectionInfo) : 
+        public OutgoingPullReplicationHandlerAsSink(ReplicationLoader parent, DocumentDatabase database, PullReplicationAsSink node, TcpConnectionInfo connectionInfo) :
             base(parent, database, node, connectionInfo)
         {
             _node = node;
